Add Frame Sync, TICP Frame Source/Sink, Add helper class for M-PSK Ambiguity Resolver and Frame Sync
Frame Sync
Add Frame synchronization, based on helper class ccsds_asm_operator.
Add GRC definition.

TICP Frame Source
Connect to a TICP Server and request Frames that are then forwarded as
a packed bytestream into GNURadio.
Add GRC definition.

TICP Frame Sink
Wait for a TICP Client to connect. If the client requests a Frame it
is taken from an asynchronous  message queue.
Add GRC definition.

M-PSK Ambiguity Resolver
Move ASM related operations to helper class ccsds_asm_operator.

DLL
Change fractional index to double precision

PLL
Remove Q-PSK rotation by 45 degrees. All constellations now have
symbols on the real axis.

+++ include/ccsds_dll_cc.h
@@ -131,7 +131,7 @@ private:
 	 *  interpolated sample lies exactly in the middle between the basepoint
 	 *  sample and the following sample.
 	 */
-	float d_mu;
+	double d_mu;
 
 	/*! \brief Storage for complex to real conversion */
 	float *d_real;
